MS3PE/requirements.txt
addict==2.4.0
albumentations==1.0.0
attr==0.3.2
attrs==22.2.0
Cython==0.29.23
easydict==1.10
einops==0.8.1
fvcore==0.1.5.post20221221
h5py==3.8.0
imageio==2.35.1
inplace_abn==1.1.0
ipython==9.4.0
matlab==0.1
matplotlib==3.7.1
mmcv==1.6.2
networkx==3.1
numpy==1.22.4
opencv_python_headless==4.10.0.84
pavi==0.0.1
Pillow==11.3.0
PyYAML==6.0.2
rfconv==0.0.2b20210406
scikit_image==0.19.3
scipy==1.16.1
skimage==0.0
timm==0.6.11
torch==1.13.1+cu116
torchvision==0.14.1+cu116
yacs==0.1.8
